backend/requirements.txt
# SoftSkills Hub — dependências do backend
# Instale com: pip install -r requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
python-dotenv>=1.0.0
slowapi>=0.1.9
